requirements.txt
Django==3.1.2
djangorestframework==3.12.1
psycopg2-binary==2.8.6

pytest~=6.2.1
django-filter~=2.4.0